@@ -1,14 +1,5 @@
-const InputView = require('./InputView');
-
 class App {
-  play() {
-    InputView.readGameStart();
-    InputView.readBridgeSize();
-  }
+  async run() {}
 }
 
-const app = new App();
-
-app.play();
-
-module.exports = App;
+export default App;
